Make prompt string object property
This change makes it possible to dynamically alter the prompt by setting
the app.prompt property.

The use case of this, for me personally, is that I have an application
that has different environments. It is important for the user to know
when they are on testing vs. production when running certain commands.

This change should be backwards-compatible and non-breaking.